+++ crates/pow/src/challenge_cache.rs
@@ -5,8 +5,8 @@ use std::time::{SystemTime, UNIX_EPOCH};
 use miden_client::account::AccountId;
 use tokio::time::{Duration, interval};
 
-use crate::pow::api_key::ApiKey;
-use crate::pow::challenge::Challenge;
+use crate::api_key::ApiKey;
+use crate::challenge::Challenge;
 
 // CHALLENGE CACHE
 // ================================================================================================