+++ crates/pow/src/lib.rs
@@ -5,29 +5,44 @@ use std::time::{SystemTime, UNIX_EPOCH};
 use miden_client::account::AccountId;
 use tokio::time::Duration;
 
-use crate::pow::api_key::ApiKey;
-use crate::pow::challenge::Challenge;
-use crate::pow::challenge_cache::ChallengeCache;
+use crate::challenge_cache::ChallengeCache;
 
-pub mod api_key;
-pub mod challenge;
+mod api_key;
+mod challenge;
 mod challenge_cache;
+
+pub use api_key::ApiKey;
+pub use challenge::Challenge;
 
 // POW
 // ================================================================================================
 
+/// Proof-of-Work implementation.
+///
+/// This struct is used to generate and validate `PoW` challenges.
 #[derive(Clone)]
-pub(crate) struct PoW {
+pub struct PoW {
+    /// The server secret used to sign and validate challenges.
     secret: [u8; 32],
+    /// The cache used to store submitted challenges.
     challenge_cache: Arc<Mutex<ChallengeCache>>,
+    /// The configuration settings.
     config: PoWConfig,
 }
 
+/// The configuration settings for `PoW`.
 #[derive(Clone)]
 pub struct PoWConfig {
+    /// The lifetime for challenges. After this time, challenges are considered expired.
     pub challenge_lifetime: Duration,
+    /// Determines how much the difficulty increases with the amount of active challenges. The
+    /// difficulty is computed as `num_active_challenges << growth_rate`.
     pub growth_rate: NonZeroUsize,
+    /// Sets the `max_target` used for challenges. The initial target (with difficulty = 1) for
+    /// challenges will be `u64::MAX >> baseline`.
     pub baseline: u8,
+    /// The interval at which the challenge cache is cleaned up. Only expired challenges are
+    /// removed during cleanup.
     pub cleanup_interval: Duration,
 }
 
@@ -51,14 +66,14 @@ impl PoW {
     }
 
     /// Generates a new challenge.
-    pub fn build_challenge(&self, request: PowRequest) -> Challenge {
+    pub fn build_challenge(&self, account_id: AccountId, api_key: ApiKey) -> Challenge {
         let current_time = SystemTime::now()
             .duration_since(UNIX_EPOCH)
             .expect("current timestamp should be greater than unix epoch")
             .as_secs();
-        let target = self.get_challenge_target(&request.api_key);
+        let target = self.get_challenge_target(&api_key);
 
-        Challenge::new(target, current_time, request.account_id, request.api_key, self.secret)
+        Challenge::new(target, current_time, account_id, api_key, self.secret)
     }
 
     /// Computes the target for a given API key by checking the amount of active challenges in the
@@ -95,7 +110,7 @@ impl PoW {
     ///
     /// # Panics
     /// Panics if the challenge cache lock is poisoned.
-    pub(crate) fn submit_challenge(
+    pub fn submit_challenge(
         &self,
         account_id: AccountId,
         api_key: &ApiKey,
@@ -135,12 +150,6 @@ impl PoW {
 
         Ok(())
     }
-}
-
-/// Validated and parsed request for the `PoW` challenge.
-pub struct PowRequest {
-    pub account_id: AccountId,
-    pub api_key: ApiKey,
 }
 
 /// `PoW` challenge related errors.
@@ -199,7 +208,7 @@ mod tests {
         let current_time = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
 
         let account_id = 0_u128.try_into().unwrap();
-        let challenge = pow.build_challenge(PowRequest { account_id, api_key: api_key.clone() });
+        let challenge = pow.build_challenge(account_id, api_key.clone());
         let nonce = find_pow_solution(&challenge, 10000).expect("Should find solution");
 
         // Submit challenge with correct nonce - should succeed
@@ -222,7 +231,7 @@ mod tests {
         let account_id = [0u8; AccountId::SERIALIZED_SIZE].try_into().unwrap();
         let current_time = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
 
-        let challenge = pow.build_challenge(PowRequest { account_id, api_key: api_key.clone() });
+        let challenge = pow.build_challenge(account_id, api_key.clone());
         let nonce = find_pow_solution(&challenge, 10000).expect("Should find solution");
 
         // Submit challenge with expired timestamp - should fail
@@ -249,7 +258,7 @@ mod tests {
         let account_id = [0u8; AccountId::SERIALIZED_SIZE].try_into().unwrap();
 
         // Solve first challenge
-        let challenge = pow.build_challenge(PowRequest { account_id, api_key: api_key.clone() });
+        let challenge = pow.build_challenge(account_id, api_key.clone());
         let nonce = find_pow_solution(&challenge, 10000).expect("Should find solution");
 
         let current_time = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
@@ -259,7 +268,7 @@ mod tests {
 
         // Try to submit second challenge - should fail because of rate limiting
         tokio::time::sleep(pow.config.cleanup_interval).await;
-        let challenge = pow.build_challenge(PowRequest { account_id, api_key: api_key.clone() });
+        let challenge = pow.build_challenge(account_id, api_key.clone());
         let nonce = find_pow_solution(&challenge, 10000).expect("Should find solution");
 
         let current_time = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
@@ -280,7 +289,7 @@ mod tests {
 
         assert_eq!(pow.get_challenge_target(&api_key), u64::MAX >> pow.config.baseline);
 
-        let challenge = pow.build_challenge(PowRequest { account_id, api_key: api_key.clone() });
+        let challenge = pow.build_challenge(account_id, api_key.clone());
         let nonce = find_pow_solution(&challenge, 10000).expect("Should find solution");
 
         pow.submit_challenge(account_id, &api_key, &challenge.encode(), nonce, current_time)
@@ -323,7 +332,7 @@ mod tests {
         assert_eq!(pow.challenge_cache.lock().unwrap().num_challenges_for_api_key(&api_key), 0);
 
         // submit second challenge - should succeed
-        let challenge = pow.build_challenge(PowRequest { account_id, api_key: api_key.clone() });
+        let challenge = pow.build_challenge(account_id, api_key.clone());
         let nonce = find_pow_solution(&challenge, 10000).expect("Should find solution");
 
         let current_time = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();